Fixes bug 1363550 - squash fs crashstorage classes
This pulls in work I did when I worked on socorro-collector. It squashes the
hierarchy of filesystem crash stores down to two: FSRaditTreeStorage and
FSTemporaryStorage.

It changes all the configuration that required other classes to use one of these
two. It removes some classes that we don't use that was using one of these
classes. It updates the tests accordingly.
