+++ socorro/collector/collector_app.py
@@ -97,8 +97,7 @@ class CollectorApp(BaseCollectorApp):
     required_config.storage.add_option(
         'crashstorage_class',
         doc='the source storage class',
-        default='socorro.external.fs.crashstorage'
-                '.FSLegacyDatedRadixTreeStorage',
+        default='socorro.external.fs.crashstorage.FSTemporaryStorage',
         from_string_converter=class_converter
     )
