+++ socorro/collector/crashmover_app.py
@@ -26,67 +26,6 @@ class CrashMoverApp(FetchTransformSaveApp):
 
 
 #==============================================================================
-class ProcessedCrashCopierApp(FetchTransformSaveWithSeparateNewCrashSourceApp):
-    app_name = 'processed_crash_copier'
-    app_version = '1.0'
-    app_description = __doc__
-
-    required_config = Namespace()
-
-    #--------------------------------------------------------------------------
-    @staticmethod
-    def get_application_defaults():
-        """this method allows an app to inject defaults into the configuration
-        that can override defaults not under the direct control of the app.
-        For example, if an app were to use a class that had a config default
-        of X and that was not appropriate as a default for this app, then
-        this method could be used to override that default.
-
-        This is a technique of getting defaults into an app that replaces
-        an older method of going to the configman option and using the
-        'set_default' method with 'force=True'"""
-
-        return {
-            'source.crashstorage_class':
-                'socorro.external.boto.crashstorage.BotoS3CrashStorage',
-            'destination.crashstorage_class':
-                'socorro.external.fs.crashstorage.TarFileCrashStore',
-            'producer_consumer.producer_consumer_class':
-                'socorro.lib.task_manager.TaskManager',
-            'producer_consumer.quit_on_empty_queue': True,
-            'new_crash_source.new_crash_source_class':
-                'socorro.processor.timemachine.PGQueryNewCrashSource'
-        }
-
-    #--------------------------------------------------------------------------
-    def _transform(self, crash_id):
-        """this default transform function only transfers raw data from the
-        source to the destination without changing the data.  While this may
-        be good enough for the raw crashmover, the processor would override
-        this method to create and save processed crashes"""
-        self.config.logger.debug('trying %s', crash_id)
-        try:
-            self.config.logger.debug('get_unredacted_processed')
-            processed_crash = self.source.get_unredacted_processed(crash_id)
-        except Exception as x:
-            self.config.logger.error(
-                "reading processed_crash: %s",
-                str(x),
-                exc_info=True
-            )
-            processed_crash = {}
-        try:
-            self.config.logger.debug('save_processed')
-            self.destination.save_processed(processed_crash)
-        except Exception as x:
-            self.config.logger.error(
-                "writing processed_crash: %s",
-                str(x),
-                exc_info=True
-            )
-
-
-#==============================================================================
 class RawAndProcessedCopierApp(FetchTransformSaveWithSeparateNewCrashSourceApp):
     """copy raw & processed crashes from a source to a destination"""
     app_name = 'raw_and_processed_crash_copier'
