+++ socorro/processor/processor_app.py
@@ -19,7 +19,7 @@ from socorro.app.fetch_transform_save_app import (
 )
 from socorro.external.crashstorage_base import CrashIDNotFound
 from socorro.lib.util import DotDict
-from socorro.external.fs.crashstorage import FSDatedPermanentStorage
+from socorro.external.fs.crashstorage import FSRadixTreeStorage
 
 
 #==============================================================================
