+++ socorro/unittest/external/fs/test_crash_data.py
@@ -82,8 +82,7 @@ class IntegrationTestCrashData(TestCase):
         mock_logging = Mock()
         required_config = Namespace()
         required_config.namespace('filesystem')
-        required_config.filesystem.filesystem_class = \
-            crashstorage.FSRadixTreeStorage
+        required_config.filesystem.filesystem_class = crashstorage.FSRadixTreeStorage
         required_config.filesystem.add_option('logger', default=mock_logging)
         config_manager = ConfigurationManager(
             [required_config],
